# docs/swagger.yaml
openapi: 3.0.3
info:
  title: SIU-2 Backend API
  version: 1.0.0
  description: Documentación OpenAPI 3.0 generada exclusivamente a partir de los routers definidos en routes/.
servers:
  - url: http://localhost:5000/api
    description: Servidor local
...
security:
  - CookieAuth: []
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
      description: Cookie con el JWT de sesión.
  schemas:
    ErrorItem:
      type: object
      required:
        - code
        - message
        - level
      properties:
        code:
          type: string
        message:
          type: string
        level:
          type: string
        description:
          type: string
    ErrorResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
...
    AuthLoginRequest:
      type: object
      required:
        - dni
        - password
      properties:
        dni:
          type: integer
        password:
          type: string
    AuthLoginResponse:
      type: object
      required:
        - mensaje
        - token
      properties:
        mensaje:
          type: string
        token:
          type: string
    AuthSignupRequest:
      type: object
      required:
        - nombre
        - apellido
        - dni
        - email
        - password
      properties:
        nombre:
          type: string
        apellido:
          type: string
        dni:
          type: integer
        email:
          type: string
        password:
          type: string
    Usuario:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        apellido:
          type: string
        email:
          type: string
        dni:
          type: integer
        es_admin:
          type: boolean
        activo:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    PerfilesResponse:
      type: object
      required:
        - perfiles
      properties:
        perfiles:
          type: array
          items:
            type: string
            enum: [admin, docente, alumno]
...
    Curso:
      type: object
      properties:
        id:
          type: integer
        materia_id:
          type: integer
        nombre:
          type: string
        anio:
          type: integer
        cuatrimestre:
          type: integer
    CursoCreateRequest:
      type: object
      required:
        - materia_id
        - nombre
        - anio
        - cuatrimestre
      properties:
        materia_id:
          type: integer
        nombre:
          type: string
        anio:
          type: integer
        cuatrimestre:
          type: integer
    CursoUpdateRequest:
      type: object
      required:
        - materia_id
        - nombre
        - anio
        - cuatrimestre
      properties:
        materia_id:
          type: integer
        nombre:
          type: string
        anio:
          type: integer
        cuatrimestre:
          type: integer
...
  responses:
    BadRequest:
      description: Solicitud inválida o datos de entrada no válidos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Falta la cookie de autenticación, el token es inválido o expiró.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: El usuario autenticado no posee el rol requerido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: El recurso ya existe o viola una restricción de unicidad.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Error inesperado del servidor.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Sesión iniciada correctamente.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/signup:
    post:
      tags: [Auth]
      summary: Registrar un nuevo usuario
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthSignupRequest'
      responses:
        '200':
          description: Usuario registrado correctamente.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [usuario]
                properties:
                  usuario:
                    $ref: '#/components/schemas/Usuario'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/me/perfiles:
    get:
      tags: [Auth]
      summary: Obtener los perfiles activos del usuario autenticado
      responses:
        '200':
          description: Perfiles recalculados desde la base de datos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerfilesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /materias/{materia_id}/cursos:
    get:
      tags: [Materias]
      summary: Listar cursos de una materia
      parameters:
        - name: materia_id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Lista paginada de cursos asociados a la materia.
          content:
            application/json:
              schema:
                type: object
                required: [cursos, total, page, page_size]
                properties:
                  cursos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Curso'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
        '204':
          description: Sin resultados.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /cursos:
    get:
      tags: [Cursos]
      summary: Listar cursos
      parameters:
        - name: materia_id
          in: query
          schema:
            type: integer
        - name: nombre
          in: query
          schema:
            type: string
        - name: anio
          in: query
          schema:
            type: integer
        - name: cuatrimestre
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Lista paginada de cursos.
          content:
            application/json:
              schema:
                type: object
                required: [cursos, total, page, page_size]
                properties:
                  cursos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Curso'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
        '204':
          description: Sin resultados.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Cursos]
      summary: Crear un curso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CursoCreateRequest'
      responses:
        '201':
          description: Curso creado correctamente.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [message, curso]
                properties:
                  message:
                    type: string
                  curso:
                    $ref: '#/components/schemas/Curso'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cursos/{curso_id}:
    parameters:
      - name: curso_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Cursos]
      summary: Obtener un curso por ID
      responses:
        '200':
          description: Curso encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Curso'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Cursos]
      summary: Reemplazar un curso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CursoUpdateRequest'
      responses:
        '200':
          description: Curso actualizado correctamente.
          content:
            application/json:
              schema:
                type: object
                required: [message, curso]
                properties:
                  message:
                    type: string
                  curso:
                    $ref: '#/components/schemas/Curso'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Cursos]
      summary: Eliminar un curso
      responses:
        '204':
          description: Curso eliminado correctamente.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /asistencia/cursos/{curso_id}/me:
    get:
      tags: [Asistencias]
      summary: Ver el porcentaje y detalle de asistencias propio del alumno logueado
      parameters:
        - name: curso_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Historial y métricas de asistencia devueltas correctamente.
          content:
            application/json:
              schema:
                type: object
                required: [curso_id, clases_totales_dictadas, presentes, ausentes, tardes, porcentaje_asistencia, condicion_asistencia, detalle]
                properties:
                  curso_id:
                    type: integer
                  clases_totales_dictadas:
                    type: integer
                  presentes:
                    type: integer
                  ausentes:
                    type: integer
                  tardes:
                    type: integer
                  porcentaje_asistencia:
                    type: number
                    format: float
                  condicion_asistencia:
                    type: string
                  detalle:
                    type: array
                    items:
                      type: object
                      properties:
                        clase_id:
                          type: integer
                        fecha:
...